+++ tests/Stubs/StringableWidget.php
@@ -9,13 +9,8 @@ use Yiisoft\Widget\Widget;
 
 final class StringableWidget extends Widget
 {
-    protected function run(): Stringable
+    public function render(): Stringable
     {
         return new StringableObject('run');
-    }
-
-    protected function afterRun(string $result): Stringable
-    {
-        return new StringableObject('after-' . parent::afterRun($result));
     }
 }